**MAJOR BUG FIX FOR DICOMIO**
In previous implementations, the dicom volume are read by `sitk.ReadImage` and `sitk.GetArrayFromImage`.
There were no orient check after the 3D-ndarray is fetched.
But the array stored in itk data is not necessarily in the same orientation.
This will introduce unexpected disalignment of data array dimension definition.

In most cases, the disalignment majorlly occurs between `L` and `R`; `P` and `A`; `I` and `S`.
When using 2D model for segmentation, such issue can be seen as an flip augmentation,
and commonly will not introduce learning problems.
But for 3D model with 3D-structure learning ability, such issue is fatal.
For a long time, I have been concentrated on model's confusion between liver and stomach,
this is quite strange, as these two organs are very easy to identify from a human view.
Now, the problem may be revealed, which is the disalignment of orientation across multi-dataset scenarios.

V1.11.9 introduces multiple IO and preprocessing fix to the above issue.
Now, I use `LPI` as the standard orientation.
The following scripts as influenced.

1. nii2mha conversion
2. mha2mha resample
3. mha2npz precrop
4. mha2npz 2D split
5. mha2npz 3D split
6. ITK Image Loading function
7. sitk_toolkit
8. Totalsegmentator-specified fixing tools
9. more are on the way.